Moved the open pending issue modification requests selector to the intake selector. Imported it and used it in the add issues page.
