+++ client/app/intake/selectors.js
@@ -41,3 +41,12 @@ export const issueCountSelector = (state) => {
 
   return selectedIssues ? selectedIssues.length : 0;
 };
+
+const selectPendingIssueModificationRequests = (state) => state.pendingIssueModificationRequests;
+
+export const getOpenPendingIssueModificationRequests = createSelector(
+  [selectPendingIssueModificationRequests],
+  (pendingIssueModificationRequests) =>
+    pendingIssueModificationRequests.filter((request) => request.status === 'assigned')
+);
+
