--- client/app/intakeEdit/selectors.js
@@ -1,9 +1,0 @@
-import { createSelector } from 'reselect';
-
-const selectPendingIssueModificationRequests = (state) => state.pendingIssueModificationRequests;
-
-export const getOpenPendingIssueModificationRequests = createSelector(
-  [selectPendingIssueModificationRequests],
-  (pendingIssueModificationRequests) =>
-    pendingIssueModificationRequests.filter((request) => request.status === 'assigned')
-);
